Fix that nasty state bug

+++ examples/local_exec/fetch.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
-wget https://releases.hashicorp.com/consul/1.9.3/consul_1.9.3_linux_amd64.zip -O /tmp/consul.zip
-cd /tmp && unzip ./consul.zip
+if [ ! -f /tmp/consul ]; then
+  wget https://releases.hashicorp.com/consul/1.9.3/consul_1.9.3_linux_amd64.zip -O /tmp/consul.zip
+  cd /tmp && unzip ./consul.zip
+fi
